:recycle:  Refactor: Dropbox 분리 in Dropdown
