+++ src/components/base/Dropdown/Style.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { font12, font14 } from "@styles/fonts";
+import { font12 } from "@styles/fonts";
 import { flexbox } from "@styles/commonStyles";
 
 interface FormProps {
@@ -36,23 +36,4 @@ export const Form = styled.form<FormProps>`
   &:hover {
     border-color: ${({ theme }) => theme.colors.skyBlue};
   }
-`;
-
-export const Dropbox = styled.div`
-  ${flexbox({ fd: "column", ai: "start" })};
-  position: absolute;
-  top: 32px;
-  left: 0;
-  min-width: 130px;
-  padding: ${({ theme }) => `${theme.gap.base * 4}px ${theme.gap.base * 3}px`};
-  border: ${({ theme }) => `1px solid ${theme.colors.gray}`};
-  border-radius: ${({ theme }) => `${theme.size.borderRadius}px`};
-  background-color: white;
-  ${font14(500)};
-  gap: ${({ theme }) => `${theme.gap.base * 2}px`};
-`;
-
-export const DropboxSet = styled.div`
-  ${flexbox({ jc: "start", ai: "center" })};
-  gap: 10px;
 `;
