+++ src/components/base/index.ts
@@ -1,1 +1,2 @@
 export { default as Dropdown } from "./Dropdown/Dropdown";
+export { default as Dropbox } from "./Dropbox/Dropbox";
